+++ CHANGELOG.md
@@ -146,3 +146,4 @@
 - 2023-10-25: Serde deals, session
 - 2023-10-26: Features deals
 - 2023-10-27: Metrics deals, qualifier, gp
+- 2023-10-28: Serde deals